+++ src/com/landawn/abacus/util/stream/ArrayCharStream.java
@@ -1239,7 +1239,7 @@ class ArrayCharStream extends AbstractCharStream {
     }
 
     @Override
-    public long sum() {
+    public int sum() {
         return sum(elements, fromIndex, toIndex);
     }
