+++ src/com/landawn/abacus/util/stream/ArrayIntStream.java
@@ -1884,7 +1884,7 @@ class ArrayIntStream extends AbstractIntStream {
     }
 
     @Override
-    public long sum() {
+    public int sum() {
         return sum(elements, fromIndex, toIndex);
     }
