+++ src/com/landawn/abacus/util/stream/ShortStream.java
@@ -534,7 +534,7 @@ public abstract class ShortStream
      */
     public abstract OptionalShort kthLargest(int k);
 
-    public abstract long sum();
+    public abstract int sum();
 
     public abstract OptionalDouble average();
